doc: visit route in swagger

+++ src/service/clients/DeleteClientService.ts
@@ -1,3 +1,5 @@
+import createError from "http-errors";
+
 /*
     [x] verificar se o cliente existe pelo ID
     [x] deletar o cliente
@@ -7,16 +9,30 @@ import prismaClient from "../../prisma";
 
 class DeleteClientService {
     async execute(id: string) {
-        //verificando se o cliente existe
-        // const clientExist = await prismaClient.client.findUnique({
-        //     where: {
-        //         id: id,
-        //     },
-        // });
+        // verificando se o cliente existe
+        const clientExist = await prismaClient.client.findUnique({
+            where: {
+                id: id,
+            },
+        });
 
-        // if (!clientExist) {
-        //     throw new Error("Cliente não existe.");
-        // }
+        if (!clientExist) {
+            throw createError(404, "Cliente não existe.");
+        }
+
+        // deletando todas as visitas
+        await prismaClient.visit.deleteMany({
+            where: {
+                client_id: id,
+            },
+        });
+
+        // deletando todos os trabalhos trabalhos
+        await prismaClient.work.deleteMany({
+            where: {
+                client_id: id,
+            },
+        });
 
         // deletando o cliente
         const deleteClient = await prismaClient.client.delete({
@@ -25,7 +41,7 @@ class DeleteClientService {
             },
         });
 
-        return deleteClient;
+        return { deleteClient };
     }
 }
